RONDB-724: Use query worker flag
This flag is a table property that disables the use of query workers.
This means that all accesses to the table will use the LDM threads and
thus the additional mutexes required is no longer necessary. This can
be useful in tables that have many partitions and where updates is a
major part of the workload.

Query workers makes key reads and scans scale on more CPUs, this comes
at a cost of additional mutexes, for some tables this is an unnecessary
cost that we can avoid.

It is possible to define it on a new table. It is possible to disable it
as an online operation. But in this case the mutex will be kept since
it is too complex to ensure that all ongoing uses of query workers have
ceased before it can be changed. However after a node restart the mutexes
are no longer required.

Enabling the use of query workers after it being disabled can be performed
as a completely online operation.

Added a bit of debugging around fragment status and a bit more to track
the use of query workers.