RONDB-724: Changed handling of table record in DBTC
Accessing the table record when sending SCAN_FRAGREQ was extremely costly
since first the tableRecord pointer had to be read, next the table pointer
had to be calculated, next the pointer had to be used read the flags field.
Finally one had to perform and AND operation between the value and a constant.

This caused as much as 20% performance regression in Sysbench OLTP RW. To avoid
this problem we write the value of use query worker into the scan record at
definition when the table record is already in CPU cache. The scan record is
also in CPU cache when used, so this means that we avoid CPU cache misses.

In addition the m_flags was removed and replaced by a set of 1-byte variable.
The size of the table record increases slightly, but should win that back
through much less code to handle the bitmaps in the m_flags field.