RONDB-934: Continous scans for normal scans by using 2 receivers and parallel ordered index scans
DBTC support for Parallel Ordered Index scan using OrderBy

Prepare DBLQH for receiving two receivers for scan fragment

Introduce new API config variable to enable/disable continous scans

Prepare NDB API for continous scans

Removed remaining use of setParallelism that impacted TTL and new features in
this patch.

Only allow continous scans for ReadCommitted scans, the reason is that otherwise
we are locking more rows and this might cause more deadlock situations and in
particular it could cause existing applications to all of a sudden start
failing, thus not backwards compatible. This risk doesn't exist for
ReadCommitted where only performance aspects will change.

Send NEXT_SCANREQ after sending SCAN_FRAGCONF if using 2 receivers.
When receiving SCAN_NEXTREQ we have two possibilities if continous scan is
happening. If it hasn't finished handling a batch to completion we will
simply set the state to IDLE and continue and if we are ready with the
batch we are ready to send SCAN_FRAGCONF and potentially continue scanning
again.

When receiving SCAN_NEXTREQ with close flag and when handling failed TC node
we could already receive this in any scan state, so we simply reset the
m_continous_scan_state to IDLE and handle the close as a usual close of a
scan.

We avoid sending scan heartbeat to DBTC while scanning where DBTC is not
expecting any signals from us.

We avoid calling closeScanLab when scan encounters error, normally this
would be ok since DBTC is ready to receive SCAN_FRAGREF from us. But here
we will postpone handling closeScanLab until we received SCAN_NEXTREQ.
We set scanCompletedStatus to ZTRUE to ensure this.

BLOBs do not use ReadCommittedFlag, this is changed after creating the
NdbReceiver objects, so continous scan should not be used in this case.
Taken care of in finaliseScan method.

Added a fair bit of debugging support.