RONDB-934: Used 4 receivers per parallelism for continous scans
Using 2 receivers was good for normal unordered scans, but for sorted index scans
one needs at least 3 receivers and to prepare for tests with higher parallelism one
might as well standardise on 4 receivers per parallelism.

It would be possible to reuse the NdbReceiver object for the parallelism, however this
would create access to the same object from two threads concurrently, so thus better
to use multiple NdbReceiver objects to avoid concurrency issues, this also secures that
no changes are required for the NdbReceiver object.