RONDB-934: Changes in DBTC for continous scan
Supporting higher parallelism in scans

DBTC needs to maintain four times as many receiver id's as the parallelism.
Thus the parallelism is a fourth of the number of receiver objects.
We have to track which receiver object to use for the moment and store
the receiver objects used by the various fragment scans.