RONDB-934: Added support for continous scans in NDB API
Supporting higher parallelism in scans

The most complex part of supporting continous scans is for sorted
scans on ordered indexes. This requires a fairly complex state
machine tracking the state of all concurrent fragment scans
and ensuring that we enter wait states when appropriate.

It is important to get this part right since much of the
performance improvements comes from avoiding too many wait states
and do as much batching as possible without overusing memory.

A lot of new debugging is added to support this.

The NdbReceiver added a new variable with its index in the array
of NdbReceiver to make it easy to correlate to its state. A new
state for NdbReceiver objects was introduced with 7 different
states.

Some minor changes to handling SCAN_TABCONF was required.

The handling of nextResult for continous scans with sorting is a
completely new method to avoid complicating the existing code.