RONDB-978: User rate limits This feature builds upon the Database Quota feature. The database quota feature used the database as the one that had rate limits and quotas. Now we can add "databases" that are users. They are treated the same as database quotas, but they can only set rate limits. Quotas (memory usage and disk usage) are still only useful for database quotas.
A transaction can only be rate limited by either database or user.
If it is rated by user, the database rates won't be checked and
the rate won't be added to the database. Thus it is designed such
that one should use either databases or users for rate limits.

TCKEYREQ and SCAN_TABREQ and TCINDXREQ have added a user id and
a user id version. These are optional parameters and not
providing will provide the same functionality as before this
feature was added.

If a user id is accepted, the user id will be sent in LQHKEYREQ
and SCAN_FRAGREQ.

The decision to include user id is done in the NDB API, a new
function setUserId is added on NdbTransaction class. If this
function is used we will translate a username to a user id and
user id version and send this in the signal to DBTC.

To speed up this translation we will store a hash table mapping
a username to a user id and user id version. The build of this
hash table will start at the first call to setUserId. When the
build is complete we will trust that the hash table is up to
date. To keep the hash table up to date we will send
a report about new users and dropped users to all connected
API nodes using the version where this feature is added.

During the build we will still send a get user signal for each
request to minimise the latency during the build process. If
another thread has already sent such a request we will wait
for this request to finish or the build phase to complete
first.

The user handling is implemented in the MySQL Server if a new
optional configuration parameter in my.cnf is set to true.
This parameter is user_rate_limits. We will use the username
of the MySQL connection as username.

The user handling is implemented in RDRS2 if a new optional
configuration parameter UserRateLimits is set to true.
In RDRS2 the user name comes from the API key, thus if no
API key is used, no user will exist.

Hopsworks uses a username which is a concatenation of project
name and username derived from API key. The RDRS2 added information
in the API key cache on username for an API key. The RDRS2 backend
concatenates the project name and username, the project name is
derived from the database name.

Internally in the data nodes the user uses database quotas. However
in the MGM server and client a new user concept is introduced which
is very similar to database quotas except that only rate limits is
supported, no quotas are supported.

Internally in data nodes it is not necessary to connect tables to
a user as with database quotas.

Internally in the RonDB data nodes the user name is stored with
a preceding $ character. This ensures that the user name can be
the same as the database name since $ is not an allowed character
in a database name.